backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
supabase
PyJWT